# openapi.yaml
openapi: 3.0.3
info:
  title: Tienda de Videojuegos DEVCO
  description: |-
    API para una tienda de videojuegos, que permite realizar las operaciones
    básicas (CRUD). 
    
    Permite gestionar los videojuegos, las ordenes y los
    usuarios de la tienda.
  contact:
    email: asalazars@devco.com.co
  version: 1.0.0
...
paths:
...
  /orden:
    get:
      tags:
        - orden
      summary: Obtiene todas las ordenes de la tienda
      description: Obtiene todas las ordenes realizadas en la tienda
      operationId: obtenerTodasOrdenes
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Orden'
      security:
        - bearerAuth: []
    post:
      tags:
        - orden
      summary: Crea una nueva orden en la tienda
      description: Genera una nueva orden en la tienda
      operationId: crearOrden
      requestBody:
        content:
          application/json:
            schema:
                type: object
                properties:
                  videojuegos_comprados:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        cantidad:
                          type: integer
                    example:
                      - id: 1
                        cantidad: 2
                      - id: 2
                        cantidad: 3
          application/xml:
            schema:
                type: object
                properties:
                  videojuegos_comprados:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        cantidad:
                          type: integer
                    example:
                      - id: 1
                        cantidad: 2
                      - id: 2
                        cantidad: 3
                xml:
                  name: orden
        required: true
      responses:
        '201':
          description: Creación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Orden'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Orden'
        '400':
          description: Datos inválidos
        '404':
          description: Videojuego dentro la orden no encontrado
  /orden/{idOrden}:
    get:
      tags:
        - orden
      summary: Obtiene una orden por id
      description: Devuelve una única orden
      operationId: obtenerOrdenPorId
      parameters:
        - name: idOrden
          in: path
          description: ID de la orden que se quiere obtener
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Orden'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Orden'
        '400':
          description: ID inválido
        '404':
          description: Orden no encontrada
      security:
        - bearerAuth: []
    delete:
      tags:
        - orden
      summary: Elimina una orden
      description: Elimina una orden
      operationId: eliminarOrden
      parameters:
        - name: idOrden
          in: path
          description: ID de la orden que se quiere obtener
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Orden'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Orden'
        '400':
          description: ID inválido
        '404':
          description: Orden no encontrada
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    Orden:
      type: object
      properties:
        id:
          type: integer
          example: 1
        videojuegos_comprados:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              cantidad:
                type: integer
          example:
            - id: 1
              cantidad: 2
            - id: 2
              cantidad: 3
        valor_total:
          type: number
          example: 100000
        cantidad:
          type: integer
          example: 5
      xml:
        name: orden
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT